// src/generated/types/createSiteRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { ProxySettingsRequest } from './proxySettingsRequest';

/**
 * Create a site request.
 */
export interface CreateSiteRequest {
  name: string;
  description?: string;
  notes?: string;
  onDemand?: boolean;
  splashtopAutoInstall?: boolean;
  proxySettings?: ProxySettingsRequest;
}